Fix pages got bundled together.
Importing URL from pages components directly is actually a bad idea in Gatsby since it tries to bundle together frequently used component.

These mistakes cause the primary bundle to go up in size from just over 80kb to over 120kb.
