+++ frontend/src/components/Page404/Page404.tsx
@@ -6,7 +6,7 @@ import Flex from "../Flex/Flex"
 import Typography from "../Typography/Typography"
 import Button from "../Button/Button"
 import Box from "../Box/Box"
-import { OBSERVE_PAGE_URL } from "../../pages/dashboard/observe"
+import { OBSERVE_PAGE_URL } from "../../routes"
 
 export const Page404: FC = () => {
   const notFound = useStaticQuery(graphql`
