+++ frontend/src/components/PageAllObservations/PageAllObservations.tsx
@@ -2,7 +2,7 @@ import React, { FC, memo, useCallback, useMemo, useState } from "react"
 import { Observation, useGetObservations } from "../../api/useGetObservations"
 import Box from "../Box/Box"
 import BackNavigation from "../BackNavigation/BackNavigation"
-import { STUDENT_DETAILS_PAGE_URL } from "../../pages/dashboard/observe/students/details"
+import { STUDENT_DETAILS_PAGE_URL } from "../../routes"
 import { categories } from "../../categories"
 import Flex from "../Flex/Flex"
 import Chip from "../Chip/Chip"
