+++ frontend/src/components/PageEditClass/PageEditClass.tsx
@@ -2,7 +2,7 @@ import React, { FC, useEffect, useRef, useState } from "react"
 import { useImmer } from "use-immer"
 import dayjs from "dayjs"
 import { navigate } from "../Link/Link"
-import { CLASS_SETTINGS_URL } from "../../pages/dashboard/settings/class"
+import { CLASS_SETTINGS_URL } from "../../routes"
 import { Box } from "../Box/Box"
 import BackNavigation from "../BackNavigation/BackNavigation"
 import { Typography } from "../Typography/Typography"
