+++ frontend/src/components/PageHome/PageHome.tsx
@@ -11,7 +11,7 @@ import { Typography } from "../Typography/Typography"
 import Card from "../Card/Card"
 import { useGetStudents } from "../../api/students/useGetStudents"
 import LoadingPlaceholder from "../LoadingPlaceholder/LoadingPlaceholder"
-import { NEW_STUDENT_URL } from "../../pages/dashboard/observe/students/new"
+import { NEW_STUDENT_URL } from "../../routes"
 import StudentPicturePlaceholder from "../StudentPicturePlaceholder/StudentPicturePlaceholder"
 
 export const PageHome: FC = () => {
