+++ frontend/src/components/PageNewClass/PageNewClass.tsx
@@ -3,7 +3,7 @@ import { useImmer } from "use-immer"
 import dayjs from "dayjs"
 import { navigate } from "../Link/Link"
 import BackNavigation from "../BackNavigation/BackNavigation"
-import { CLASS_SETTINGS_URL } from "../../pages/dashboard/settings/class"
+import { CLASS_SETTINGS_URL } from "../../routes"
 import { Box } from "../Box/Box"
 import Input from "../Input/Input"
 import Flex from "../Flex/Flex"
