+++ frontend/src/components/PagePickGuardian/PagePickGuardian.tsx
@@ -2,7 +2,7 @@ import React, { FC, useState } from "react"
 import { navigate } from "../Link/Link"
 import Box from "../Box/Box"
 import BackNavigation from "../BackNavigation/BackNavigation"
-import { NEW_STUDENT_URL } from "../../pages/dashboard/observe/students/new"
+import { NEW_STUDENT_URL } from "../../routes"
 import Input from "../Input/Input"
 import Select from "../Select/Select"
 import { GuardianRelationship } from "../../api/students/usePostNewStudent"
