+++ frontend/src/components/PageSettings/PageSettings.tsx
@@ -14,7 +14,7 @@ import { ReactComponent as FlipIcon } from "../../icons/flip.svg"
 import Card from "../Card/Card"
 import { useGetSchool } from "../../api/useGetSchool"
 import LoadingPlaceholder from "../LoadingPlaceholder/LoadingPlaceholder"
-import { CLASS_SETTINGS_URL } from "../../pages/dashboard/settings/class"
+import { CLASS_SETTINGS_URL } from "../../routes"
 
 export const PageSettings: FC = () => {
   const schoolDetail = useGetSchool()
