+++ frontend/src/components/PageStudentDetails/PageStudentDetails.tsx
@@ -20,7 +20,7 @@ import StudentProgressSummaryCard from "../StudentProgressSummaryCard/StudentPro
 import { ReactComponent as NextIcon } from "../../icons/next-arrow.svg"
 import { ReactComponent as PrevIcon } from "../../icons/arrow-back.svg"
 import { ReactComponent as PlusIcon } from "../../icons/plus.svg"
-import { ALL_OBSERVATIONS_PAGE_URL } from "../../pages/dashboard/observe/students/observations/all"
+import { ALL_OBSERVATIONS_PAGE_URL } from "../../routes"
 import dayjs from "../../dayjs"
 
 interface Props {
