+++ frontend/src/pages/dashboard/observe/students/details.tsx
@@ -4,9 +4,6 @@ import queryString from "query-string"
 import PageStudentDetails from "../../../../components/PageStudentDetails/PageStudentDetails"
 import SEO from "../../../../components/seo"
 
-export const STUDENT_DETAILS_PAGE_URL = (studentId: string): string =>
-  `/dashboard/observe/students/details?id=${studentId}`
-
 const StudentDetails: FC<PageRendererProps> = ({ location }) => {
   const query = queryString.parse(location.search)
   let id: string
