+++ frontend/src/pages/dashboard/observe/students/guardians/pick.tsx
@@ -1,8 +1,6 @@
 import React, { FC } from "react"
 import PagePickGuardian from "../../../../../components/PagePickGuardian/PagePickGuardian"
 
-export const PICK_GUARDIAN_URL = "/dashboard/observe/students/guardians/pick"
-
 const PickGuardian: FC = () => <PagePickGuardian />
 
 export default PickGuardian
