+++ frontend/src/pages/dashboard/observe/students/new.tsx
@@ -3,8 +3,6 @@ import { PageRendererProps } from "gatsby"
 import SEO from "../../../../components/seo"
 import PageNewStudent from "../../../../components/PageNewStudent/PageNewStudent"
 
-export const NEW_STUDENT_URL = "/dashboard/observe/students/new"
-
 const NewStudent: FC<PageRendererProps> = ({ location }) => {
   return (
     <>
