+++ frontend/src/pages/dashboard/observe/students/observations/all.tsx
@@ -3,9 +3,6 @@ import SEO from "../../../../../components/seo"
 import PageAllObservations from "../../../../../components/PageAllObservations/PageAllObservations"
 import { useQueryString } from "../../../../../hooks/useQueryString"
 
-export const ALL_OBSERVATIONS_PAGE_URL = (studentId: string): string =>
-  `/dashboard/observe/students/observations/all?studentId=${studentId}`
-
 const AllObservations: FC = () => {
   const studentId = useQueryString("studentId")
