+++ frontend/src/pages/dashboard/settings/class/edit.tsx
@@ -3,9 +3,6 @@ import SEO from "../../../../components/seo"
 import PageEditClass from "../../../../components/PageEditClass/PageEditClass"
 import { useQueryString } from "../../../../hooks/useQueryString"
 
-export const EDIT_CLASS_URL = (classId: string): string =>
-  `/dashboard/settings/class/edit?classId=${classId}`
-
 const EditClass: FC = () => {
   const classId = useQueryString("classId")
