+++ frontend/src/pages/dashboard/settings/class/index.tsx
@@ -2,8 +2,6 @@ import React, { FC } from "react"
 import SEO from "../../../../components/seo"
 import PageClassSettings from "../../../../components/PageClassSettings/PageClassSettings"
 
-export const CLASS_SETTINGS_URL = "/dashboard/settings/class"
-
 const Class: FC = () => {
   return (
     <>
