+++ frontend/src/pages/dashboard/settings/class/new.tsx
@@ -2,8 +2,6 @@ import React, { FC } from "react"
 import SEO from "../../../../components/seo"
 import PageNewClass from "../../../../components/PageNewClass/PageNewClass"
 
-export const NEW_CLASS_URL = "/dashboard/settings/class/new"
-
 const NewClass: FC = () => {
   return (
     <>
