+++ frontend/src/pages/dashboard/settings/curriculum/subjects/edit.tsx
@@ -3,9 +3,6 @@ import SEO from "../../../../../components/seo"
 import { useQueryString } from "../../../../../hooks/useQueryString"
 import PageEditSubject from "../../../../../components/PageEditSubject/PageEditSubject"
 
-export const EDIT_SUBJECT_URL = (areaId: string, subjectId: string): string =>
-  `/dashboard/settings/curriculum/subjects/edit?subjectId=${subjectId}&areaId=${areaId}`
-
 const EditSubject: FC = () => {
   const subjectId = useQueryString("subjectId")
   const areaId = useQueryString("areaId")
