+++ frontend/src/pages/dashboard/settings/index.tsx
@@ -2,8 +2,6 @@ import React, { FC } from "react"
 import PageSettings from "../../../components/PageSettings/PageSettings"
 import SEO from "../../../components/seo"
 
-export const SETTINGS_URL = "/dashboard/settings"
-
 const Settings: FC = () => {
   return (
     <>
